Fixed a potential compatibility problem.
The plain int type is only guaranteed to be at least 16 bits, and
yet the flags variable was expecting at least 22 bits. This turns
out to be true for x86 and x86-64 systems, but to ensure compatibility,
the variable has been changed to an unsigned long int. Also added
the "L" suffix to flags that were more than 16 bits.

+++ src/structs.h
@@ -60,7 +60,7 @@ struct object {
 	int collectType; // What the object is carrying
 	int collectValue; // What it is worth
 
-	int flags; // Various flags for an object
+	unsigned long int flags; // Various flags for an object
 
 	float x, y, dx, dy;
